+++ src/main/java/ar/com/gvallerino/xMenML/dto/DnaStatsResponse.java
@@ -2,37 +2,37 @@ package ar.com.gvallerino.xMenML.dto;
 
 public class DnaStatsResponse {
 	
-	private long count_mutant_dna;
-	private long count_human_dna;
-	private long ratio;
+	private double count_mutant_dna;
+	private double count_human_dna;
+	private double ratio;
 	
-	public DnaStatsResponse (long mutant, long dna, long ratio) {
+	public DnaStatsResponse (double mutant, double dna, double ratio) {
 		this.count_mutant_dna = mutant;
 		this.count_human_dna = dna;
 		this.ratio = ratio;
 	}
 	
-	public long getCount_mutant_dna() {
+	public double getCount_mutant_dna() {
 		return count_mutant_dna;
 	}
 	
-	public void setCount_mutant_dna(long count_mutant_dna) {
+	public void setCount_mutant_dna(double count_mutant_dna) {
 		this.count_mutant_dna = count_mutant_dna;
 	}
 	
-	public long getCount_human_dna() {
+	public double getCount_human_dna() {
 		return count_human_dna;
 	}
 	
-	public void setCount_human_dna(long count_human_dna) {
+	public void setCount_human_dna(double count_human_dna) {
 		this.count_human_dna = count_human_dna;
 	}
 	
-	public long getRatio() {
+	public double getRatio() {
 		return ratio;
 	}
 	
-	public void setRatio(long ratio) {
+	public void setRatio(double ratio) {
 		this.ratio = ratio;
 	}
